Cleanup. Extracted some classes to separate files. Removed unnecessary suppression in TileData.cs. Fixed tooltip in Options.
