+++ UoFiddler.Controls/UserControls/ItemShowAlternativeControl.cs
@@ -587,7 +587,7 @@ namespace UoFiddler.Controls.UserControls
                 return;
             }
 
-            Animdata.Data info = Animdata.GetAnimData(id);
+            Animdata.AnimdataEntry info = Animdata.GetAnimData(id);
             if (info != null)
             {
                 DetailTextBox.AppendText(
